Refactor; better names and plugged in iabd reasons
