+++ app/uk/gov/hmrc/tai/util/yourTaxFreeAmount/YourTaxFreeAmount.scala
@@ -21,7 +21,7 @@ import play.api.i18n.Messages
 import uk.gov.hmrc.play.language.LanguageUtils.Dates
 import uk.gov.hmrc.tai.model.TaxYear
 import uk.gov.hmrc.tai.model.domain.calculation.CodingComponent
-import uk.gov.hmrc.tai.service.yourTaxFreeAmount.YourTaxFreeAmountComparison
+import uk.gov.hmrc.tai.service.YourTaxFreeAmountComparison
 import uk.gov.hmrc.tai.util.{TaxAccountCalculator, TaxAccountCalculatorImpl, TaxYearRangeUtil}
 
 trait YourTaxFreeAmount {
