+++ test/uk/gov/hmrc/tai/service/yourTaxFreeAmount/IabdTaxCodeChangeServiceSpec.scala
@@ -24,19 +24,20 @@ import play.api.i18n.{I18nSupport, MessagesApi}
 import uk.gov.hmrc.tai.model.domain._
 import uk.gov.hmrc.tai.util.yourTaxFreeAmount._
 
-class YourTaxFreeAmountComparisonServiceSpec extends PlaySpec with MockitoSugar with FakeTaiPlayApplication with I18nSupport {
+class IabdTaxCodeChangeServiceSpec extends PlaySpec with MockitoSugar with FakeTaiPlayApplication with I18nSupport {
 
   private val taxFreeInfo = TaxFreeInfo("12-12-2015", 2000, 1000)
   private val jobExpensesIncrease = CodingComponentPair(JobExpenses, Some(2), Some(50), Some(100))
   private val carBenefitIncrease = CodingComponentPair(CarBenefit, Some(1), Some(1000), Some(2000))
   private val taxCodeChange = mock[TaxCodeChange]
   implicit val messagesApi = app.injector.instanceOf[MessagesApi]
+  val service = new IabdTaxCodeChangeService
 
-  "taxCodeChangeReasons method" must {
+  "iabdReasons method" must {
     "have no reasons" when {
       "there are no allowances and deductions" in {
         val pairs = AllowancesAndDeductionPairs(Seq.empty, Seq.empty)
-        val reasons = YourTaxFreeAmountComparisonService.taxCodeChangeReasons(pairs)
+        val reasons = service.iabdReasons(pairs)
 
         reasons mustBe Seq.empty
       }
@@ -45,7 +46,7 @@ class YourTaxFreeAmountComparisonServiceSpec extends PlaySpec with MockitoSugar 
         val noPreviousAmount = CodingComponentPair(JobExpenses, None, None, Some(123))
         val pairs = AllowancesAndDeductionPairs(Seq(noPreviousAmount), Seq.empty)
 
-        val reasons = YourTaxFreeAmountComparisonService.taxCodeChangeReasons(pairs)
+        val reasons = service.iabdReasons(pairs)
 
         reasons mustBe Seq.empty
       }
@@ -54,7 +55,7 @@ class YourTaxFreeAmountComparisonServiceSpec extends PlaySpec with MockitoSugar 
         val noCurrentAmount = CodingComponentPair(JobExpenses, None, Some(123), None)
         val pairs = AllowancesAndDeductionPairs(Seq(noCurrentAmount), Seq.empty)
 
-        val reasons = YourTaxFreeAmountComparisonService.taxCodeChangeReasons(pairs)
+        val reasons = service.iabdReasons(pairs)
 
         reasons mustBe Seq.empty
       }
@@ -62,7 +63,7 @@ class YourTaxFreeAmountComparisonServiceSpec extends PlaySpec with MockitoSugar 
 
     "give multiple reasons for a tax code change" in {
       val pairs = AllowancesAndDeductionPairs(Seq(jobExpensesIncrease), Seq(carBenefitIncrease))
-      val reasons = YourTaxFreeAmountComparisonService.taxCodeChangeReasons(pairs)
+      val reasons = service.iabdReasons(pairs)
 
       reasons mustBe Seq(
         "Your Job expenses has been updated",
@@ -73,7 +74,7 @@ class YourTaxFreeAmountComparisonServiceSpec extends PlaySpec with MockitoSugar 
     "show benefit changes once" in {
       val pairs = AllowancesAndDeductionPairs(Seq.empty, Seq(carBenefitIncrease, carBenefitIncrease))
 
-      val reasons = YourTaxFreeAmountComparisonService.taxCodeChangeReasons(pairs)
+      val reasons = service.iabdReasons(pairs)
 
       reasons mustBe Seq("Your Car benefit has been updated")
     }
@@ -87,7 +88,7 @@ class YourTaxFreeAmountComparisonServiceSpec extends PlaySpec with MockitoSugar 
         val pairs = AllowancesAndDeductionPairs(Seq.empty,
           Seq(carBenefitIncrease, jobExpensesIncrease, giftSharesCharity, flatRateJobExpense, lossRelief))
 
-        val reasons = YourTaxFreeAmountComparisonService.taxCodeChangeReasons(pairs)
+        val reasons = service.iabdReasons(pairs)
 
         reasons mustBe Seq(messagesApi("taxCode.change.yourTaxCodeChanged.paragraph"))
       }
