+++ test/uk/gov/hmrc/tai/util/yourTaxFreeAmount/YourTaxFreeAmountSpec.scala
@@ -23,7 +23,7 @@ import org.scalatestplus.play.PlaySpec
 import play.api.i18n.Messages
 import uk.gov.hmrc.play.language.LanguageUtils.Dates
 import uk.gov.hmrc.tai.model.TaxYear
-import uk.gov.hmrc.tai.service.yourTaxFreeAmount.YourTaxFreeAmountComparison
+import uk.gov.hmrc.tai.service.YourTaxFreeAmountComparison
 import uk.gov.hmrc.tai.util.TaxYearRangeUtil
